# internal/workspace/api/oas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"fmt"
	"time"

	"github.com/google/uuid"
)

func (s *ErrorStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

type CookieAuth struct {
	APIKey string
	Roles  []string
}

// GetAPIKey returns the value of APIKey.
func (s *CookieAuth) GetAPIKey() string {
	return s.APIKey
}

// GetRoles returns the value of Roles.
func (s *CookieAuth) GetRoles() []string {
	return s.Roles
}

// SetAPIKey sets the value of APIKey.
func (s *CookieAuth) SetAPIKey(val string) {
	s.APIKey = val
}

// SetRoles sets the value of Roles.
func (s *CookieAuth) SetRoles(val []string) {
	s.Roles = val
}

// Ref: #/components/schemas/Error
type Error struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

// GetCode returns the value of Code.
func (s *Error) GetCode() int {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *Error) GetMessage() string {
	return s.Message
}

// SetCode sets the value of Code.
func (s *Error) SetCode(val int) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *Error) SetMessage(val string) {
	s.Message = val
}

// ErrorStatusCode wraps Error with StatusCode.
type ErrorStatusCode struct {
	StatusCode int
	Response   Error
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorStatusCode) GetResponse() Error {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorStatusCode) SetResponse(val Error) {
	s.Response = val
}

type ID uuid.UUID

// Ref: #/components/schemas/Project
type Project struct {
	ID          ID          `json:"id"`
	WorkspaceId ID          `json:"workspaceId"`
	Name        ProjectName `json:"name"`
	CreatedAt   time.Time   `json:"createdAt"`
}

// GetID returns the value of ID.
func (s *Project) GetID() ID {
	return s.ID
}

// GetWorkspaceId returns the value of WorkspaceId.
func (s *Project) GetWorkspaceId() ID {
	return s.WorkspaceId
}

// GetName returns the value of Name.
func (s *Project) GetName() ProjectName {
	return s.Name
}

// GetCreatedAt returns the value of CreatedAt.
func (s *Project) GetCreatedAt() time.Time {
	return s.CreatedAt
}

// SetID sets the value of ID.
func (s *Project) SetID(val ID) {
	s.ID = val
}

// SetWorkspaceId sets the value of WorkspaceId.
func (s *Project) SetWorkspaceId(val ID) {
	s.WorkspaceId = val
}

// SetName sets the value of Name.
func (s *Project) SetName(val ProjectName) {
	s.Name = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *Project) SetCreatedAt(val time.Time) {
	s.CreatedAt = val
}

type ProjectName string

type WorkspacesWorkspaceIdProjectsPostReq struct {
	Name ProjectName `json:"name"`
}

// GetName returns the value of Name.
func (s *WorkspacesWorkspaceIdProjectsPostReq) GetName() ProjectName {
	return s.Name
}

// SetName sets the value of Name.
func (s *WorkspacesWorkspaceIdProjectsPostReq) SetName(val ProjectName) {
	s.Name = val
}

// WorkspacesWorkspaceIdProjectsProjectIdDeleteNoContent is response for WorkspacesWorkspaceIdProjectsProjectIdDelete operation.
type WorkspacesWorkspaceIdProjectsProjectIdDeleteNoContent struct{}

type WorkspacesWorkspaceIdProjectsProjectIdPutReq struct {
	Name ProjectName `json:"name"`
}

// GetName returns the value of Name.
func (s *WorkspacesWorkspaceIdProjectsProjectIdPutReq) GetName() ProjectName {
	return s.Name
}

// SetName sets the value of Name.
func (s *WorkspacesWorkspaceIdProjectsProjectIdPutReq) SetName(val ProjectName) {
	s.Name = val
}